add BrowserStack chrome extension
BrowserStack is a handy Chrome extension for web developers which lets you test your work on every possible on-demand browser and OS. This plugin provides real devices to test on and not emulators/simulators. It launches new test sessions on real browsers with a couple of clicks. You can also initiate up to 12 browsers for effortless access, switch between browsers, and spend less time on the overall testing process, using the Chrome Extension.